[androsdk-881] Move FeatureType from period to common module
