+++ core/src/androidTest/java/org/hisp/dhis/android/core/enrollment/CreateEnrollmentUtils.java
@@ -34,9 +34,9 @@ import androidx.annotation.NonNull;
 
 import org.hisp.dhis.android.core.common.BaseDataModel;
 import org.hisp.dhis.android.core.common.BaseIdentifiableObjectModel;
+import org.hisp.dhis.android.core.common.FeatureType;
 import org.hisp.dhis.android.core.common.State;
 import org.hisp.dhis.android.core.enrollment.internal.EnrollmentFields;
-import org.hisp.dhis.android.core.period.FeatureType;
 
 public class CreateEnrollmentUtils {
     private static final String UID = "test_enrollment";
