+++ core/src/androidTest/java/org/hisp/dhis/android/core/program/ProgramStageModelShould.java
@@ -31,9 +31,11 @@ package org.hisp.dhis.android.core.program;
 import android.content.ContentValues;
 import android.database.MatrixCursor;
 
+import androidx.test.runner.AndroidJUnit4;
+
 import org.hisp.dhis.android.core.common.BaseIdentifiableObject;
+import org.hisp.dhis.android.core.common.FeatureType;
 import org.hisp.dhis.android.core.common.FormType;
-import org.hisp.dhis.android.core.period.FeatureType;
 import org.hisp.dhis.android.core.period.PeriodType;
 import org.hisp.dhis.android.core.program.ProgramStageModel.Columns;
 import org.hisp.dhis.android.core.utils.Utils;
@@ -41,8 +43,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import java.util.Date;
-
-import androidx.test.runner.AndroidJUnit4;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.hisp.dhis.android.core.AndroidTestUtils.toBoolean;
