+++ core/src/main/java/org/hisp/dhis/android/core/arch/helpers/CoordinateHelper.java
@@ -31,8 +31,8 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import org.hisp.dhis.android.core.common.Coordinates;
+import org.hisp.dhis.android.core.common.FeatureType;
 import org.hisp.dhis.android.core.common.Geometry;
-import org.hisp.dhis.android.core.period.FeatureType;
 
 import java.util.Arrays;
 import java.util.List;
