+++ core/src/main/java/org/hisp/dhis/android/core/common/FeatureType.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.period;
+package org.hisp.dhis.android.core.common;
 
 public enum FeatureType {
