+++ core/src/main/java/org/hisp/dhis/android/core/data/serialization/GeometryTypeSerializer.java
@@ -32,7 +32,7 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 
-import org.hisp.dhis.android.core.period.FeatureType;
+import org.hisp.dhis.android.core.common.FeatureType;
 
 import java.io.IOException;
