+++ core/src/sharedTest/java/org/hisp/dhis/android/core/data/program/ProgramSamples.java
@@ -31,7 +31,7 @@ package org.hisp.dhis.android.core.data.program;
 import org.hisp.dhis.android.core.category.CategoryCombo;
 import org.hisp.dhis.android.core.common.Access;
 import org.hisp.dhis.android.core.common.DataAccess;
-import org.hisp.dhis.android.core.period.FeatureType;
+import org.hisp.dhis.android.core.common.FeatureType;
 import org.hisp.dhis.android.core.program.Program;
 import org.hisp.dhis.android.core.program.ProgramType;
 import org.hisp.dhis.android.core.relationship.RelationshipType;
