+++ core/src/test/java/org/hisp/dhis/android/core/arch/helpers/CoordinateHelperShould.java
@@ -30,8 +30,8 @@ package org.hisp.dhis.android.core.arch.helpers;
 
 import org.assertj.core.util.Lists;
 import org.hisp.dhis.android.core.common.Coordinates;
+import org.hisp.dhis.android.core.common.FeatureType;
 import org.hisp.dhis.android.core.common.Geometry;
-import org.hisp.dhis.android.core.period.FeatureType;
 import org.junit.Test;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
