+++ core/src/test/java/org/hisp/dhis/android/core/enrollment/EnrollmentShould.java
@@ -31,8 +31,8 @@ package org.hisp.dhis.android.core.enrollment;
 import org.hisp.dhis.android.core.common.BaseIdentifiableObject;
 import org.hisp.dhis.android.core.common.BaseObjectShould;
 import org.hisp.dhis.android.core.common.Coordinates;
+import org.hisp.dhis.android.core.common.FeatureType;
 import org.hisp.dhis.android.core.common.ObjectShould;
-import org.hisp.dhis.android.core.period.FeatureType;
 import org.junit.Test;
 
 import java.io.IOException;
