+++ core/src/test/java/org/hisp/dhis/android/core/event/EventShould.java
@@ -30,8 +30,8 @@ package org.hisp.dhis.android.core.event;
 
 import org.hisp.dhis.android.core.common.BaseIdentifiableObject;
 import org.hisp.dhis.android.core.common.BaseObjectShould;
+import org.hisp.dhis.android.core.common.FeatureType;
 import org.hisp.dhis.android.core.common.ObjectShould;
-import org.hisp.dhis.android.core.period.FeatureType;
 import org.junit.Test;
 
 import java.io.IOException;
